+++ test/integration/scriptEndpointTracesTest.js
@@ -27,7 +27,7 @@ describe('ScriptEndpointTrace', function() {
       return connection.Script.please().create({
         instanceName: instanceName,
         label: scriptEndpointName,
-        runtime_name: 'python',
+        runtime_name: 'python_library_v4.2',
         source: 'print "x"'
       });
     }).then((script) => {